Move old documentation to a temporary folder, waiting for full rewrite
In the mean time, most of the documentation for [eleventy-plugin-images-responsiver](https://github.com/nhoizey/eleventy-plugin-images-responsiver), the Eleventy plugin that uses it, should be enough to understand how it works.

"Just" replace any Markdown with the HTML it is transformed into, and use the same configuration to get the same result.